Clean up. Move build steps to project file
