--- .teamcity/Ribasim_Windows/buildTypes/Windows_BuildRibasim.kt
@@ -1,11 +1,0 @@
-package Ribasim_Windows.buildTypes
-
-import Templates.BuildWindows
-import Templates.GithubCommitStatusIntegration
-import Templates.WindowsAgent
-import jetbrains.buildServer.configs.kotlin.BuildType
-
-object Windows_BuildRibasim : BuildType({
-    templates(WindowsAgent, GithubCommitStatusIntegration, BuildWindows)
-    name = "Build Ribasim"
-})
